requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-storages>=1.14.2
boto3>=1.34.0
python-dotenv>=1.0.0
celery>=5.3.6
redis>=5.0.1
sendgrid>=6.11.0
requests==2.31.0
django-jazzmin==2.6.0 